refactor(web): extract judge LabelFilter into pages/judge/LabelFilter.tsx (PRD #1049)
